Refactored the hp registaration api and added a new api to submit the registartion request to initiate the workflow.

+++ src/main/java/in/gov/abdm/nmr/repository/SuperSpecialityRepository.java
@@ -7,7 +7,7 @@ import in.gov.abdm.nmr.entity.SuperSpeciality;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface SuperSpecialityRepository extends JpaRepository<SuperSpeciality, Long> {
+public interface SuperSpecialityRepository extends JpaRepository<SuperSpeciality, BigInteger> {
 
     @Query(value = "SELECT name, id FROM super_speciality", nativeQuery = true)
     List<SuperSpeciality> getSpeciality();
